backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlmodel>=0.0.22
python-dotenv>=1.0.0
anthropic>=0.42.0
openai>=1.60.0
google-genai>=1.0.0
python-multipart>=0.0.9
